Simplify expression and just run the first day of the month
It doesn't seem like GH actions is actually evaluating the day of the week part (it run today, for example).

The recommended site https://crontab.guru/#0_2_1-7_*_6 shows the expression as:

> “At 02:00 on every day-of-month from 1 through 7 and on Saturday.”

The AND on Saturday seems to imply that it will run all week in addition to Saturday, so it's not what we want.